+++ .eslintrc.json
@@ -1,0 +1,32 @@
+{
+	"root": true,
+	"env": {
+		"node": true,
+		"browser": true
+	},
+	"extends": [
+		"eslint:recommended",
+		"plugin:astro/recommended",
+		"plugin:mdx/recommended",
+		"plugin:@typescript-eslint/recommended"
+	],
+	"parser": "@typescript-eslint/parser",
+	"plugins": ["@typescript-eslint"],
+	"settings": {
+    "mdx/code-blocks": true
+	},
+	"overrides": [
+		{
+			"files": ["*.astro"],
+			"parser": "astro-eslint-parser",
+			"parserOptions": {
+				"parser": "@typescript-eslint/parser",
+				"extraFileExtensions": [".astro"]
+			},
+			"rules": {
+				// override/add rules settings here, such as:
+				// "astro/no-set-html-directive": "error"
+			}
+		}
+	]
+}
