+++ .vscode/extensions.json
@@ -1,7 +1,8 @@
 {
   "recommendations": [
     "astro-build.astro-vscode",
-    "phoenisx.cssvar"
+    "phoenisx.cssvar",
+    "unifiedjs.vscode-mdx"
   ],
   "unwantedRecommendations": []
 }
