+++ package.json
@@ -5,17 +5,33 @@
   "private": true,
   "devDependencies": {
     "@netlify/plugin-lighthouse": "4.0.3",
-    "open-props": "^1.4.25",
+    "@typescript-eslint/eslint-plugin": "5.48.0",
+    "@typescript-eslint/parser": "5.48.0",
+    "eslint": "8.31.0",
+    "eslint-plugin-astro": "0.21.1",
+    "eslint-plugin-jsx-a11y": "6.6.1",
+    "eslint-plugin-mdx": "2.0.5",
+    "npm-run-all": "4.1.5",
+    "open-props": "^1.5.1",
+    "postcss-custom-media": "8.0.2",
     "postcss-jit-props": "^1.0.7",
-    "postcss-preset-env": "^7.8.3"
+    "postcss-nesting": "10.2.0",
+    "prettier": "2.8.2",
+    "prettier-plugin-astro": "0.7.2"
   },
   "dependencies": {
-    "@astrojs/mdx": "^0.11.6",
+    "@astrojs/mdx": "^0.12.0",
     "@astrojs/rss": "^1.0.3",
     "@astrojs/sitemap": "^1.0.0",
-    "astro": "^1.6.10"
+    "astro": "^1.6.12"
   },
   "scripts": {
+    "lint:format": "prettier --write src",
+    "lint:syntax": "eslint . --ext .js,.mjs,.cjs,.ts,.astro,.mdx",
+    "lint": "run-s lint:*",
+    "check:format": "prettier --check src",
+    "check:types": "tsc --noEmit",
+    "checks": "run-s check:*",
     "dev": "astro dev",
     "start": "astro dev",
     "build": "astro build",
@@ -24,6 +40,6 @@
   },
   "volta": {
     "node": "18.12.1",
-    "npm": "8.19.3"
+    "npm": "9.1.2"
   }
 }
