+++ src/components/base-head.astro
@@ -4,9 +4,9 @@
 import "../styles/global.css";
 
 export interface Props {
-  title: string;
-  description: string;
-  image?: string;
+	title: string;
+	description: string;
+	image?: string;
 }
 
 const { title, description, image = "/placeholder-social.jpg" } = Astro.props;
@@ -19,10 +19,10 @@ const { title, description, image = "/placeholder-social.jpg" } = Astro.props;
 <link rel="me" href="https://toot.cafe/@oliverturner" />
 <link rel="icon" type="image/svg+xml" href="/favicon.svg" />
 <link
-  rel="alternate"
-  type="application/rss+xml"
-  title="RSS Feed for oliverturner.cloud"
-  href="/rss.xml"
+	rel="alternate"
+	type="application/rss+xml"
+	title="RSS Feed for oliverturner.cloud"
+	href="/rss.xml"
 />
 
 <meta name="generator" content={Astro.generator} />
