+++ src/components/header-link.astro
@@ -1,5 +1,5 @@
 ---
-export interface Props extends astroHTML.JSX.AnchorHTMLAttributes {}
+// export interface Props extends astroHTML.JSX.AnchorHTMLAttributes {}
 
 const { href, class: className, ...props } = Astro.props;
 
@@ -8,17 +8,17 @@ const isActive = href === pathname || href === pathname.replace(/\/$/, "");
 ---
 
 <a href={href} class:list={[className, { active: isActive }]} {...props}>
-  <slot />
+	<slot />
 </a>
 
 <style>
-  a {
-    display: inline-block;
-    text-decoration: none;
+	a {
+		display: inline-block;
+		text-decoration: none;
 
-    &.active {
-      font-weight: bolder;
-      text-decoration: underline;
-    }
-  }
+		&.active {
+			font-weight: bolder;
+			text-decoration: underline;
+		}
+	}
 </style>
