+++ src/components/icon-link.astro
@@ -3,20 +3,20 @@ const { icon, href } = Astro.props;
 ---
 
 <a class="iconlink" {href} target="_blank">
-  <svg><use xlink:href={`/spritesheet.svg#icon-${icon}`}></use></svg>
-  <span><slot /></span>
+	<svg><use xlink:href={`/spritesheet.svg#icon-${icon}`}></use></svg>
+	<span><slot /></span>
 </a>
 
 <style>
-  .iconlink {
-    width: 1.5rem;
-    aspect-ratio: 1;
-    overflow: hidden;
+	.iconlink {
+		width: 1.5rem;
+		aspect-ratio: 1;
+		overflow: hidden;
 
-    & svg {
-      width: 100%;
-      height: 100%;
-      fill: var(--text-mid);
-    }
-  }
+		& svg {
+			width: 100%;
+			height: 100%;
+			fill: var(--text-mid);
+		}
+	}
 </style>
