+++ src/components/site-header.astro
@@ -1,52 +1,51 @@
 ---
-import Link from "./header-link.astro";
 import { SITE_TITLE } from "../config";
 ---
 
 <header>
-  <h2>
-    <a href="/">{SITE_TITLE}</a>
-  </h2>
+	<h2>
+		<a href="/">{SITE_TITLE}</a>
+	</h2>
 </header>
 
 <style>
-  header {
-    display: grid;
-    text-align: center;
+	header {
+		display: grid;
+		text-align: center;
 
-    @media (min-width: 640px) {
-      display: flex;
-      justify-content: space-between;
-      align-items: center;
-      gap: var(--step--1);
+		@media (--mq-medium) {
+			display: flex;
+			justify-content: space-between;
+			align-items: center;
+			gap: var(--step--1);
 
-      text-align: left;
-    }
-  }
+			text-align: left;
+		}
+	}
 
-  h2 {
-    margin: 0;
-    padding: var(--step--2) var(--step--1);
-    border-bottom-right-radius: 0.25rem;
+	h2 {
+		margin: 0;
+		padding: var(--step--2) var(--step--1);
+		border-bottom-right-radius: 0.25rem;
 
-    line-height: 1;
-    font-weight: 200;
-    letter-spacing: 2px;
-    background: var(--site-header-bg);
-    color: var(--site-header-text);
+		line-height: 1;
+		font-weight: 200;
+		letter-spacing: 2px;
+		background: var(--site-header-bg);
+		color: var(--site-header-text);
 
-    & a {
-      color: inherit;
-      text-decoration: none;
-    }
-  }
+		& a {
+			color: inherit;
+			text-decoration: none;
+		}
+	}
 
-  nav {
-    display: flex;
-    gap: var(--step-0);
-    justify-content: center;
+	nav {
+		display: flex;
+		gap: var(--step-0);
+		justify-content: center;
 
-    padding: var(--step--2) var(--step--1);
-    color: var(--text-link);
-  }
+		padding: var(--step--2) var(--step--1);
+		color: var(--text-link);
+	}
 </style>
