+++ src/components/site-links.astro
@@ -1,0 +1,5 @@
+<a href="/">Home</a>
+<a href="/projects">Projects</a>
+<a href="/articles">Articles</a>
+<a href="/blog">Blog posts</a>
+<a href="/about">About</a>
