--- src/components/site-nav.astro
@@ -1,23 +1,0 @@
-<nav>
-  <a href="/">Home</a>
-  <a href="/projects">Projects</a>
-  <a href="/articles">Articles</a>
-  <a href="/blog">Blog posts</a>
-  <a href="/about">About</a>
-</nav>
-
-<style>
-  nav {
-    display: grid;
-    align-content: start;
-
-    height: 100%;
-    padding: var(--nav-padding);
-    background-color: var(--nav-bg-colour);
-    color: var(--nav-text-colour);
-  }
-
-  a {
-    text-decoration: none;
-  }
-</style>
