+++ src/layouts/page-blogpost.astro
@@ -3,81 +3,75 @@ import ContentFooter from "../components/content-footer.astro";
 import Page from "./page.astro";
 
 export interface Props {
-  content: {
-    title: string;
-    description: string;
-    pubDate?: string;
-    updatedDate?: string;
-    heroImage?: string;
-  };
+	content: {
+		title: string;
+		description: string;
+		pubDate?: string;
+		updatedDate?: string;
+		heroImage?: string;
+	};
 }
 
 const {
-  content: { title, description, pubDate, updatedDate, heroImage },
+	content: { title, description, pubDate, updatedDate, heroImage },
 } = Astro.props;
 ---
 
 <Page title={title} description={description}>
-  {
-    heroImage && (
-      <img
-        class="hero"
-        width={720}
-        height={360}
-        src={heroImage}
-        alt=""
-      />
-    )
-  }
-  <header class="header">
-    <h1 class="title">{title}</h1>
-    {pubDate && <time>{pubDate}</time>}
-    {
-      updatedDate && (
-        <p>
-          Last updated on <time>{updatedDate}</time>
-        </p>
-      )
-    }
-  </header>
-  <hr />
-  <div class="copy">
-    <slot />
-  </div>
-  <ContentFooter />
+	{
+		heroImage && (
+			<img class="hero" width={720} height={360} src={heroImage} alt="" />
+		)
+	}
+	<header class="header">
+		<h1 class="title">{title}</h1>
+		{pubDate && <time>{pubDate}</time>}
+		{
+			updatedDate && (
+				<p>
+					Last updated: <time>{updatedDate}</time>
+				</p>
+			)
+		}
+	</header>
+	<hr />
+	<div class="copy">
+		<slot />
+	</div>
+	<ContentFooter />
 </Page>
 
 <style>
-  .hero {
-    display: block;
-    width: 100%;
-    max-height: 360px;
-    object-fit: contain;
-  }
+	.hero {
+		display: block;
+		width: 100%;
+		max-height: 360px;
+		object-fit: contain;
+	}
 
-  .header {
-    padding: var(--step-0);
+	.header {
+		padding: var(--step-0);
 
-    & p {
-      margin: 0;
-      font-size: var(--step--1);
-    }
-  }
+		& p {
+			margin: 0;
+			font-size: var(--step--1);
+		}
+	}
 
-  .title {
-    font-size: 2em;
-    margin: 0;
-  }
-  hr {
-    margin: 0;
-    border-top: 1px solid var(--gray-4);
-  }
+	.title {
+		font-size: 2em;
+		margin: 0;
+	}
+	hr {
+		margin: 0;
+		border-top: 1px solid var(--gray-4);
+	}
 
-  .copy {
-    padding: var(--step-2) var(--step-3);
-  }
+	.copy {
+		padding: var(--step-2) var(--step-3);
+	}
 
-  nav {
-    /* background-color: var(--bg-transparent-dark); */
-  }
+	nav {
+		/* background-color: var(--bg-transparent-dark); */
+	}
 </style>
