+++ src/pages/about.md
@@ -1,11 +1,11 @@
 ---
 layout: "../layouts/page-blogpost.astro"
 title: "About Me"
-description: "Lorem ipsum dolor sit amet"
-updatedDate: "20th November 2022"
+description: "A "
+updatedDate: "27th November 2022"
 ---
 
-Hi! I'm Oliver. I don't know how meaningful I consider the term "full-stack" to be, but I do know that I love CSS and web standards just as much as I enjoy designing Cloud architecture, building scalable APIs, or optimising server throughput. I got into software development (after deciding I didn't want to be a barrister) in about 2007 because I wanted to help make the web more accessible; that effort remains a large part of why I still love what I do.
+Hi! I'm Oliver. I don't know how meaningful I consider the term "full-stack" to be, but I do know that I love CSS and web standards just as much as I enjoy designing Cloud architecture, building scalable APIs, and thinking about systems and observability. I got into software development (after deciding I didn't want to be a barrister) in about 2007 because I wanted to help make the web more accessible; fifteen years later that aspiration continues to guide & inspire.
 
 I work at [The Financial Times](https://ft.com) on the Advertising and Privacy team, where I'm having a blast building a new ad platform for the FT Group. You can read more about that journey in this series of articles on [_Project Gwyneth_](/articles/project-gwyneth).
 
@@ -15,4 +15,4 @@ In my free time I co-organise [London CSS](https://londoncss.dev), coach at [Cod
 
 My wife and I live in North London together with Lesley the fox. You can follow Lesley's adventures on [Instagram](https://instagram.com/harringayfoxfam) (warning: danger of cute overload).
 
-If you would like to natter, I can be found on [Mastodon](https://toot.cafe/oliverturner). My resume is available on [LinkedIn](https://www.linkedin.com/in/oliver-turner-68846313/).
+If you would like to natter, I can be found on [Mastodon](https://toot.cafe/oliverturner). If you're interested working together, my resumé is available on [LinkedIn](https://www.linkedin.com/in/oliver-turner-68846313/).
