+++ src/pages/blog.astro
@@ -4,36 +4,36 @@ import { SITE_TITLE } from "../config";
 
 const title = `Blog - ${SITE_TITLE}`;
 const description =
-  "Musings on the web, design, and development. Occasionally, foxes.";
+	"Musings on the web, design, and development. Occasionally, foxes.";
 
 // Use Astro.glob() to fetch all posts, and then sort them by date.
 const posts = (await Astro.glob("./blog/*.{md,mdx}")).sort(
-  (a, b) =>
-    new Date(b.frontmatter.pubDate).valueOf() -
-    new Date(a.frontmatter.pubDate).valueOf()
+	(a, b) =>
+		new Date(b.frontmatter.pubDate).valueOf() -
+		new Date(a.frontmatter.pubDate).valueOf()
 );
 ---
 
 <Page title={title} description={description}>
-  <h1>Week notes</h1>
-  <ul>
-    {
-      posts.map((post) => (
-        <li>
-          <time datetime={post.frontmatter.pubDate}>
-            {new Date(post.frontmatter.pubDate).toLocaleDateString("en-us", {
-              year: "numeric",
-              month: "short",
-              day: "numeric",
-            })}
-          </time>
-          <a href={post.url}>{post.frontmatter.title}</a>
-        </li>
-      ))
-    }
-  </ul>
+	<h1>Week notes</h1>
+	<ul>
+		{
+			posts.map((post) => (
+				<li>
+					<time datetime={post.frontmatter.pubDate}>
+						{new Date(post.frontmatter.pubDate).toLocaleDateString("en-us", {
+							year: "numeric",
+							month: "short",
+							day: "numeric",
+						})}
+					</time>
+					<a href={post.url}>{post.frontmatter.title}</a>
+				</li>
+			))
+		}
+	</ul>
 </Page>
 
 <style>
-  /* placeholder */
+	/* placeholder */
 </style>
