+++ src/pages/index.astro
@@ -2,17 +2,17 @@
 import Page from "../layouts/page.astro";
 
 const props = {
-  title: "Welcome",
-  description: "Oliver Turner's personal website",
-}
+	title: "Welcome",
+	description: "Oliver Turner's personal website",
+};
 ---
 
 <Page {...props}>
-  <div></div>
+	<div></div>
 </Page>
 
 <style>
-  :root {
-    --content-bg: transparent;
-  }
+	:root {
+		--content-bg: transparent;
+	}
 </style>
