+++ src/pages/projects.astro
@@ -1,22 +1,73 @@
 ---
-import Page from "../layouts/page.astro";
+interface Page {
+	frontmatter: {
+		title: string;
+		url: string;
+	};
+	url: string;
+}
+
+import Layout from "../layouts/page.astro";
+import projects from "../data/projects.json";
 import { SITE_TITLE } from "../config";
 
 const title = `Projects - ${SITE_TITLE}`;
 const description = "Deeper dives into web development";
 
 // Use Astro.glob() to fetch all posts, and then sort them by date.
-const posts = (await Astro.glob("./projects/*.{md,mdx}")).sort(
-  (a, b) =>
-    new Date(b.frontmatter.pubDate).valueOf() -
-    new Date(a.frontmatter.pubDate).valueOf()
-);
+const pages: Page[] = await Astro.glob("./projects/*.{md,mdx}");
+
+const data = [];
+for (const page of pages) {
+	console.log(page.frontmatter.title, page.url);
+	const pageUrl = page.url;
+	const { title, description } = page.frontmatter;
+	data.push({ title, description, pageUrl, ...projects[page.url] });
+}
+
+console.log(data);
 ---
 
-<Page title={title} description={description}>
-  <h1>Projects</h1>
-</Page>
+<Layout title={title} description={description}>
+	<h1>Projects</h1>
+	<div class="projects">
+		{
+			data.map((project) => (
+				<article class="project">
+					<a href={project.pageUrl}>
+						<h2>{project.title}</h2>
+						<p>{project.description}</p>
+						<p>Read more...</p>
+					</a>
+					<div>
+						<a href={project.repo_url}>GitHub</a>
+						<a href="https://app.netlify.com/sites/jeffsum/deploys">
+							<img
+								src={`https://api.netlify.com/api/v1/badges/${project.platform.id}/deploy-status`}
+								alt={`Build status of ${project.title}`}
+							/>
+						</a>
+						<a href={project.app_url}>
+							<p>Screenshot</p>
+						</a>
+					</div>
+				</article>
+			))
+		}
+	</div>
+</Layout>
 
 <style>
-  /* placeholder */
+	:root {
+		--content-bg: transparent;
+	}
+	.projects {
+		display: grid;
+		grid-template-columns: repeat(auto-fit, minmax(300px, 1fr));
+		grid-gap: 1rem;
+	}
+
+	.project {
+		background-color: var(--gray-1);
+	}
 </style>
