+++ src/pages/projects/home.md
@@ -1,0 +1,8 @@
+---
+layout: "../../layouts/page-blogpost.astro"
+title: "This website"
+description: "Lorem ipsum dolor sit amet"
+updatedDate: "19th November 2022"
+---
+
+## Home
