+++ src/pages/projects/ikon.md
@@ -1,8 +1,8 @@
 ---
 layout: "../../layouts/page-blogpost.astro"
-title: "Coming soon"
+title: "Ikon"
 description: "Lorem ipsum dolor sit amet"
 updatedDate: "19th November 2022"
 ---
 
-## Coming soon
+## Ikon
