+++ src/pages/projects/paintlets.md
@@ -1,0 +1,8 @@
+---
+layout: "../../layouts/page-blogpost.astro"
+title: "Paintlets"
+description: "Lorem ipsum dolor sit amet"
+updatedDate: "19th November 2022"
+---
+
+## Paintlets
