+++ src/pages/rss.xml.js
@@ -1,10 +1,10 @@
-import rss from '@astrojs/rss';
-import { SITE_TITLE, SITE_DESCRIPTION } from '../config';
+import rss from "@astrojs/rss";
+import { SITE_TITLE, SITE_DESCRIPTION } from "../config";
 
 export const get = () =>
 	rss({
 		title: SITE_TITLE,
 		description: SITE_DESCRIPTION,
 		site: import.meta.env.SITE,
-		items: import.meta.glob('./blog/**/*.{md,mdx}'),
+		items: import.meta.glob("./blog/**/*.{md,mdx}"),
 	});
