+++ src/styles/global.css
@@ -7,139 +7,131 @@
 /* @link https://utopia.fyi/type/calculator?c=320,16,1.2,1140,16,1.25,5,2,&s=0.75|0.5|0.25,1.5|2|3|4|6,s-l */
 
 :root {
-  --step--2: clamp(0.64rem, calc(0.72rem + -0.11vw), 0.69rem);
-  --step--1: clamp(0.8rem, calc(0.85rem + -0.06vw), 0.83rem);
-  --step-0: clamp(1rem, calc(1rem + 0vw), 1rem);
-  --step-1: clamp(1.2rem, calc(1.18rem + 0.1vw), 1.25rem);
-  --step-2: clamp(1.44rem, calc(1.39rem + 0.24vw), 1.56rem);
-  --step-3: clamp(1.73rem, calc(1.64rem + 0.44vw), 1.95rem);
-  --step-4: clamp(2.07rem, calc(1.93rem + 0.72vw), 2.44rem);
-  --step-5: clamp(2.49rem, calc(2.27rem + 1.1vw), 3.05rem);
+	--step--2: clamp(0.64rem, calc(0.72rem + -0.11vw), 0.69rem);
+	--step--1: clamp(0.8rem, calc(0.85rem + -0.06vw), 0.83rem);
+	--step-0: clamp(1rem, calc(1rem + 0vw), 1rem);
+	--step-1: clamp(1.2rem, calc(1.18rem + 0.1vw), 1.25rem);
+	--step-2: clamp(1.44rem, calc(1.39rem + 0.24vw), 1.56rem);
+	--step-3: clamp(1.73rem, calc(1.64rem + 0.44vw), 1.95rem);
+	--step-4: clamp(2.07rem, calc(1.93rem + 0.72vw), 2.44rem);
+	--step-5: clamp(2.49rem, calc(2.27rem + 1.1vw), 3.05rem);
 
-  --text-light: var(--gray-1);
-  --text-mid: var(--gray-6);
-  --text-normal: var(--gray-8);
-  --text-strong: var(--gray-9);
-  --text-link: var(--blue-6);
+	--text-light: var(--gray-1);
+	--text-mid: var(--gray-6);
+	--text-normal: var(--gray-8);
+	--text-strong: var(--gray-9);
+	--text-link: var(--blue-6);
 
-  --site-header-bg: var(--bg-transparent-dark);
-  --site-header-text: var(--text-light);
+	--site-header-bg: var(--bg-transparent-dark);
+	--site-header-text: var(--text-light);
 
-  --site-footer-border: 1px solid var(--gray-4);
+	--site-footer-border: 1px solid var(--gray-4);
 
-  --bg-transparent-dark: #000b;
-  --tint-bg: #0c0e0d;
-  --tint-bg-start: #000;
-  --tint-bg-end: darkslateblue;
+	--bg-transparent-dark: #0008;
+	--tint-bg: #0c0e0d;
+	--tint-bg-start: #000;
+	--tint-bg-end: darkslateblue;
 
+	--content-bg: var(--gray-1);
 
-  --content-width: 65ch;
-  --content-bg: var(--gray-1);
-
-  --nav-width: 40ch;
-  --nav-padding: var(--step-1);
-  --nav-bg-colour: var(--bg-transparent-dark);
-  --nav-text-colour: var(--text-light);
+	--nav-padding: var(--step-1);
+	--nav-bg-colour: var(--bg-transparent-dark);
+	--nav-text-colour: var(--text-light);
 }
 
 *,
 *::before,
 *::after {
-  box-sizing: border-box;
+	box-sizing: border-box;
 }
 
 :where(body) {
-  margin: 0;
-  font-family: system-ui, Verdana, sans-serif;
-  word-wrap: break-word;
-  overflow-wrap: break-word;
-  line-height: 1.5;
-  background-repeat: no-repeat;
+	margin: 0;
+	font-family: system-ui, Verdana, sans-serif;
+	word-wrap: break-word;
+	overflow-wrap: break-word;
+	line-height: 1.5;
 
-  background-image: url("/bg.avif"),
-    linear-gradient(180deg, var(--tint-bg-start), var(--tint-bg-end));
-  background-position: center;
-  background-size: cover;
+	background-color: var(--tint-bg);
+	background-image: linear-gradient(
+		180deg,
+		var(--tint-bg-start),
+		var(--tint-bg-end)
+	);
 
-  @media (min-width: 900px) {
-    background-size: auto;
-    background-position: center top;
-  }
-
-  background-color: var(--tint-bg);
-  color: var(--text-normal);
+	color: var(--text-normal);
 }
 
 :where(h1, h2, h3, h4, h5, h6, strong, b) {
-  color: var(--text-strong);
+	color: var(--text-strong);
 }
 
 :where(main) {
-  max-width: var(--content-width);
-  margin: auto;
+	max-width: var(--width-content);
+	margin: auto;
 }
 
 :where(a) {
-  color: inherit;
+	color: inherit;
 }
 
 textarea {
-  width: 100%;
-  font-size: 16px;
+	width: 100%;
+	font-size: 16px;
 }
 
 input {
-  font-size: 16px;
+	font-size: 16px;
 }
 
 content {
-  line-height: 1.6;
+	line-height: 1.6;
 }
 
 table {
-  width: 100%;
+	width: 100%;
 }
 
 :where(img) {
-  max-width: 100%;
-  height: auto;
+	max-width: 100%;
+	height: auto;
 }
 
 code {
-  padding: 2px 5px;
-  background-color: #f2f2f2;
+	padding: 2px 5px;
+	background-color: #f2f2f2;
 }
 
 pre {
-  padding: 1rem;
+	padding: 1rem;
 
-  & > code {
-    all: unset;
-  }
+	& > code {
+		all: unset;
+	}
 }
 
 blockquote {
-  border: 1px solid #999;
-  color: #222;
-  padding: 2px 0px 2px 20px;
-  margin: 0px;
-  font-style: italic;
+	border: 1px solid #999;
+	color: #222;
+	padding: 2px 0px 2px 20px;
+	margin: 0px;
+	font-style: italic;
 }
 
 :where(.copy) {
-  padding: var(--step-0);
+	padding: var(--step-0);
 
-  & > * {
-    margin: 0;
-  }
+	& > * {
+		margin: 0;
+	}
 
-  & > * + * {
-    margin-top: var(--step-1);
-  }
+	& > * + * {
+		margin-top: var(--step-1);
+	}
 
-  & img {
-    display: block;
-    margin: var(--step-2) auto;
-    max-height: 30ch;
-  }
+	& img {
+		display: block;
+		margin: var(--step-2) auto;
+		max-height: 30ch;
+	}
 }
