+++ tsconfig.json
@@ -1,3 +1,8 @@
 {
-  "extends": "astro/tsconfigs/strict"
+	"extends": "astro/tsconfigs/strict",
+	"compilerOptions": {
+		"types": ["vite/client"],
+	},
+	"include": ["src/**/*"],
+	"exclude": ["node_modules", ".git", ".astro", "dist"]
 }
